third_party/WebKit/LayoutTests/bluetooth/characteristic/getDescriptors/gen-characteristic-is-removed-with-uuid.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let val = new Uint8Array([1]);
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['generic_access']}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('generic_access'))
    .then(service => service.getCharacteristic('gap.device_name'))
    .then(characteristic => {
      return setBluetoothFakeAdapter('MissingCharacteristicHeartRateAdapter')
        .then(() => assert_promise_rejects_with_message(
          characteristic.getDescriptors(user_description.name),
          new DOMException(
            'GATT Characteristic no longer exists.',
            'InvalidStateError'),
          'Characteristic got removed.'));
    });
}, 'Characteristic gets removed. Reject with InvalidStateError.');


</script>
